chore: move components that were bunched with routes to components directory chore: re-fix vite.config.ts
