+++ src/components/forum/forumCategory.component.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { styled } from "@linaria/react";
-import Widget from "../../components/common/widgets/widget.component";
+import Widget from "../common/widgets/widget.component";
 import { Board } from "../../types/forum";
 import { Link } from "react-router-dom";
