+++ src/vite.config.ts
@@ -1,6 +1,6 @@
 import { defineConfig } from "vite";
 import reactRefresh from "@vitejs/plugin-react-refresh";
-import linaria from "@linaria/vite";
+import linaria from "@wyw-in-js/vite";
 
 // https://vitejs.dev/config/
 export default defineConfig({
